Moved most widgets to components package
